exceptions

+++ src/main/java/com/example/BackFinal/service/PedidoService.java
@@ -1,5 +1,6 @@
 package com.example.BackFinal.service;
 
+import com.example.BackFinal.exceptions.ResourceNotFoundException;
 import com.example.BackFinal.model.Pedido;
 import com.example.BackFinal.model.Producto;
 import com.example.BackFinal.repository.PedidoRepository;
@@ -20,7 +21,9 @@ public class PedidoService {
     public Pedido GuardarPedido(Pedido pedido) {
         return pedidoRepository.save(pedido);
     }
-    public void eliminar(Integer id) {
+    public void eliminar(Integer id) throws ResourceNotFoundException {
+        if (buscar(id).isEmpty())
+            throw new ResourceNotFoundException("No existe el pedido con el id: " + id);
         pedidoRepository.deleteById(id);
     }
     public Optional<Pedido> buscar(Integer id) {
